requirements.txt
streamlit>=1.28.0
mistralai>=1.0.0
chromadb>=0.4.0
pandas>=2.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0


